requirements.txt
fastapi
uvicorn
python-multipart
python-chess
torch
pandas
numpy
